clean up some role impl
